+++ R/plottingFunctions.R
@@ -74,10 +74,9 @@ plotMsigWordcloud <-
 #'
 #' @param ig an igraph object, containing a network of gene set overlaps
 #'   computed using [computeMsigNetwork()].
-#' @param markGroups a named list, of character vectors or numeric indices
-#'   specifying node groupings. Each element of the list represent a group and
-#'   contains a character vector with node names. Up to 12 groups can be
-#'   visualised in the plot.
+#' @param markGroups a named list, of character vectors. Each element of the
+#'   list represent a group and contains a character vector with node names. Up
+#'   to 12 groups can be visualised in the plot.
 #' @param genesetStat a named numeric, statistic to project onto the nodes.
 #'   These could be p-values, log fold-changes or gene set score from a
 #'   singscore-based analysis.
@@ -97,7 +96,10 @@ plotMsigWordcloud <-
 #' data(hgsc)
 #' ovlap <- computeMsigOverlap(hgsc)
 #' ig <- computeMsigNetwork(ovlap, hgsc)
-#' groups <- list('g1' = c(1, 9), 'g2' = c(5, 6))
+#' groups <- list(
+#'   'g1' = c("HALLMARK_HYPOXIA", "HALLMARK_GLYCOLYSIS"),
+#'   'g2' = c("HALLMARK_INTERFERON_GAMMA_RESPONSE")
+#' )
 #'
 #' plotMsigNetwork(ig, markGroups = groups)
 #' 
@@ -177,6 +179,10 @@ plotMsigNetwork <-
     
     #mark groups
     if (!is.null(markGroups)) {
+      if (!all(unlist(markGroups) %in% p1$data$name)) {
+        stop('Invalid node names found in markGroups.')
+      }
+      
       #add gene set counts for each group
       names(markGroups) = sapply(names(markGroups), function(x) {
         paste0(x, ' (n = ', length(markGroups[[x]]), ')')
@@ -184,7 +190,7 @@ plotMsigNetwork <-
       
       #complex hull for groups
       hulldf = plyr::ldply(markGroups, function(x) {
-        df = p1$data[x, ]
+        df = p1$data[p1$data$name %in% x, ]
         df = df[grDevices::chull(df$x, df$y), ]
       }, .id = 'NodeGroup')