+++ NEWS.md
@@ -1,6 +1,8 @@
 # version 1.2.0
 
 * Added weighting to text-mining analysis. Word clouds can now incorporate statistics.
+* Network plotting now performed using ggraph.
+* Removal of excessive warnings produced when performing text-mining.
 
 # version 1.0.0